Tryed to implemented Cache organizations.
I tryed to make it so first it askes the local indexeddb, for an organization
and if it's not there then it askes Alice for it, and if she returns an
organization then that one is saved to the local indexed db.
But i ran into a problem about saving the organization, and i suspect that
it's because Map.toString() do not return the map in JSON format and therefore
indexeddb can't parse it correctly and gives a exception.